requirements.txt
azure-functions
azure-mgmt-resource
azure-mgmt-datafactory
azure-keyvault-secrets
python-dotenv
azure-identity